fix: Refatorando para diminuir repetição de codigo
